scripts/mqtt_test/pyproject.toml
[project]
name = "mqtt-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["paho-mqtt>=2.1.0"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "T201", "BLE001"]
